Move Spree module definition to lib/spree
spree/core defines Spree::Core, doesn't also need to define
whats in the Spree module
