+++ core/lib/spree.rb
@@ -1,4 +1,28 @@
 module Spree
+  mattr_accessor :user_class
+
+  def self.user_class
+    if @@user_class.is_a?(Class)
+      raise "Spree.user_class MUST be a String or Symbol object, not a Class object."
+    elsif @@user_class.is_a?(String) || @@user_class.is_a?(Symbol)
+      @@user_class.to_s.constantize
+    end
+  end
+
+  # Used to configure Solidus.
+  #
+  # Example:
+  #
+  #   Spree.config do |config|
+  #     config.track_inventory_levels = false
+  #   end
+  #
+  # This method is defined within the core gem on purpose.
+  # Some people may only wish to use the Core part of Solidus.
+  def self.config(&_block)
+    yield(Spree::Config)
+  end
+
   def self.table_name_prefix
     'spree_'
   end
